+++ NEWS.md
@@ -1,5 +1,12 @@
 # User visible changes in `StructuredArrays` package
 
+## Version 0.2.5
+
+- New fast uniform arrays built by `FastUniformArray(val,dims)` which are
+  immutable uniform arrays but with the elements value being part of the
+  signature so that this value is known at compile time.
+
+
 ## Version 0.2.3
 
 - Use package `ArrayTools`.